e2e_open_data_pipeline/requirements.txt
# Dependencias para Airflow (Extracción y Transformación)
pandas==2.1.4
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Dependencias para Streamlit (Visualización)
streamlit==1.31.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.2
